01-Data-Storage-and-Batch-Pipelines/04-ETL-and-introduction-to-Airflow/Track-long/pyproject.toml
[tool.poetry]
name = "4-ETL-and-introduction-to-Airflow/Track-long"
version = "0.1.0"
description = ""
authors = ["S. Saunier <seb@lewagon.org>"]

[tool.poetry.dependencies]
python = "^3.8.12"
apache-airflow = "2.2.5"
apache-airflow-providers-postgres = "4.1.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-mock = "3.2.0"
apache-airflow-backport-providers-sqlite = "2021.3.3"
pytest-responses = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
